Create _dev solutions and projects.
There is a need to having 2 solutions: one that only has the samples and shall be used by the end-users to build samples. This solution references the nuget packages for iothub_client. This solution is iothub_client.sln. A second solution names iothub_client_dev.sln shall not use the iothub client nuget packages, but rather build the iothub client. This _dev.sln shall also have all the tests and requirement docs in it.
The same logic of creating a serializer_dev applies to the serializer project.