Reported props&send event callback fcn per send, not per handle (#1100)
See #1068 for details.  Basically an application that has different callback functions for sending events and/or reporting properties cannot have a callback function per send.  We're today storing the callback handle at the handle level.

This fixes so we tag the context of the send with the callback function and use that during the dispatcher routine instead of the global.