Rename MAX interfaces to NUM interfaces in samples
The #define for number of interfaces had MAX in its name.  This incorrectly implied there was some sort of maximum number of interfaces enforced by policy.

In reality, this is just the # of interfaces that this particular EXE happens to implement.

Rename "MAX" to "NUM" so that this is clearer.