Public preview pnp block on destroy (#1173)
If an application calls `DigitalTwin_InterfaceClient_Destroy`, we need to block return of the destroy until all callbacks have completed processing.  It doesn't add that much code and protects device application from having to synchronize this on their own.

Fix is analogous to what the `client_core` layer already does with device level handles.  Record that we're in a callback and then if destroy happens at same time, poll on destroy.  

Because we have only a single dispatcher thread to process callbacks on, a simple bool to monitor callbacks being processed is sufficient complexity here.