Fix the cancellation of the longhaul jobs in the gate for Windows platforms
The longhaul tests on Windows have been cancelled by devops after 12h, because not all Windows LH tests run in parallel.
There are 10 tests, but on our scripts we restrict the max number of parallel tests to 8.
We could fix this in different ways, but this is the cheapest and most straightforward.
Also, on Linux we run up to 16 tests in parallel, so increasing by that much on Windows should be acceptable.