Always delete xio in MQTT cleanup [LTS Security Fix]
[Port of commit # 4ac2326654a77a94361bb65fc9dbb162b690b12f into lts_07_2019 branch]

Fixes #924

At issue is that there exists a path on MQTT destruction where `xio_destroy` is not invoked, causing a memory leak:
* MQTT enters a disconnected state and stays there.  E.G. `currPacketState` = `DISCONNECT_TYPE`.
* App calls through to a `IoTHubTransport_MQTT_Common_Destroy`.
* The `DisconnectFromClient` call we're relying on will NOT free up the xio, because it's in `DISCONNECT_TYPE`.

The fix here is to make `free_transport_handle_data` (which is effectively the destructor of this handle) *always* destroy the xio.  This gets us an unconditional free so we're not dependent on other state to free up data.

I did further code review around this and saw that `conn_attempted` was set to TRUE but never reset on the xio going down.  Since the `conn_attempted` feels implicitly joined to a given xio instance, make it so that destroying an xio will reset.

Also adding a UT here.  I was able to repro @maskimthedog issue pre-fix (Valgrind showed a leak running this regression specific UT) and verified via Valgrind that it's addressed post fix.