Support for DPS enrollment groups
It's unclear how this sample works with DPS enrollment groups.  Added a comment to let users know that a device key must be generated from the DPS master key.  Note that other code samples take this into account and ask the user for a DPS Enrollment Group key and/or Individual enrollment key.