Remove vestigial code around PnP interface registration (#1496)
The function `DT_InterfaceList_CreateRegistrationMessage` is no longer called by any code anymore.  The background is when we had the "magic" telemetry message to register PnP devices, we needed this.  But now that this is encoded in the MQTT CONNECT packet, it's not needed.

This change removes `DT_InterfaceList_CreateRegistrationMessage` and the fairly extensive set of helpers that it calls.  Also fixes up helper functions where there were options that only `DT_InterfaceList_CreateRegistrationMessage` paths invoked.  And of course fixes UT.