Move from urn's to dtmi (#1500)
Interfaces in DTDLv2 are no longer prefixed with `urn:` but instead follow a `dtmi:` pattern.  Furthermore, the separator between the name and the version is now a `;'.

This change:
* Renames `urn` => `dtmi` in samples, documentation, UT, etc.
* Changes the parser to validate `dtmi` and componentNames are valid, following DTDLv2 rules.

Typically we don't validate on client itself and let server fail policy issues.  But a developer getting this wrong creates very unpleasant debugging experience (I've seen this in practice).  

The application will send the invalid data to the server and the server will fail by killing the TCP connection.  The underlying IoTHub SDK will assume this TCP Connection dropping is because of a network issue, not caller's fault, and kick in its retry logic.  The application developer will understandably also think something is wrong with their network or IoTHub and not think to look at how they formatted a string.  

So... we do a local check first to provide meaningful error information in the developer inner loop.