Samples: fixed and unified comment related to setting the trusted certificate
The comment was ambiguous in some files, saying that, setting the Trusted
Certificate is only necessary on systems "with" and "without" built in
certificate store.
As already stated in other files, it is only necessary on systems without
certificate store (for example: on small embedded systems).

The comment has been then unified all over the files containing it, for
consistency and uniformity.