Document number is treated as a double, not a long (#1531)
We're discouraging (if not outright removing) the use of longs from DTDL.  The issue is that Node cannot easily handle numbers > 2^52, whereas a long implies 2^64.  This will cause interop problems for customers using Node.

We're changing the DeviceInfo DTDL to use doubles instead of long.  While it may seem over-precise to specify a KB as a double instead of just as an int, our options of long we don't like and just moving it to a 32bit int would limit the maximum storage reportable on a device to 2TB which clearly isn't enough.

There's not really protocol changes in what the SDK is sending to support this.  Our comments should be correct, though.