Add ModelID parameter to IoTHub Client (MQTT only) (#1552)
In order to support applications building PnP / DigitalTwin clients on the `master` IoTHub API, we need to let them specify their modelId for the MQTT CONNECT packet, which lives in the userName field.  

This is done via a `DeviceClient_SetOption(handle, OPTION_MODEL_ID, "dtmi:...");`.  It must be specified prior to the first CONNECT, since trying to add this or allow it to be changed later will considerably complicate the state management.  The only time a modelId will change on a given device/module identity in practice is during firmware updates in any case, which implies a new handle in any case.

The modelID parsing feature requires an API version which is still in preview and not deployed worldwide.  To account for this, we will use a preview API version if modelId has been specified and otherwise use the existing model ID.  This API version split is not long-term desirable; #1547 tracks fixing.

Note that this change is MQTT(_WS) only, per decision to only support C device SDK with PnP over MQTT(_WS).