Fix cross-comp CI pipeline (ARM & MIPS32) to build against target branch instead of main
The ARM and the MIPS32 cross-compilation jobs in .vsts-ci.yml will always target the main branch,
ignoring the branch a given PR is based on and letting potential issues sip through.
This change fixes the ARM and MIPS32 jobs so they run against whatever branch the pipeline is running
against.