fix: compare result to CONNECTION_AUTHENTICATED
This appears to be a theoretical bug. The sample code is comparing two values
of disparate enum types. The enumerators that are being used in the comparison
have the same values (and so the comparison behavior is valid), but other uses of
this enumeration type suggest that this was a typo.